Reduce the number of C++ compiler warnings in AMEGIC/AMEGIC-generated sources
(cherry picked from commit a6ebec233fac9ff13622ba7fdc240b3e35e44ecf)

+++ AddOns/Higgs/PS3/C3_5.C
@@ -36,7 +36,7 @@ extern "C" Single_Channel * Getter_C3_5(int nin,int nout,Flavour* fl,Integration
 void C3_5::GeneratePoint(Vec4D * p,Cut_Data * cuts,double * _ran)
 {
   double *ran = p_vegas->GeneratePoint(_ran);
-  for(int i=0;i<m_rannum;i++) p_rans[i]=ran[i];
+  for(size_t i=0;i<m_rannum;i++) p_rans[i]=ran[i];
   Vec4D p234=p[0]+p[1];
   double s234_max = p234.Abs2();
   double s3 = p_ms[3];
